MAIN: Move pdata tests to module, lessen scope of AsyncPDataWriter::new
